requirements.txt
flask
google-generativeai
python-dotenv
streamlit